Dev bruce

+++ user/models.py
@@ -32,6 +32,7 @@ class Host(models.Model):
     h_id = models.AutoField(primary_key=True)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     phone = models.CharField(max_length=20, unique=True)
-    rating = models.PositiveIntegerField(default=0)
+#    rating = models.PositiveIntegerField(default=0)
+    rating = models.DecimalField(max_digits=2, decimal_places=1, default=0.0)
 
     c_time = models.DateTimeField(auto_now_add=True)
